Moved the services to service provider plugins
