+++ docs/conf.py
@@ -46,6 +46,7 @@ shutil.move('../ChangeLog', '_static/ChangeLog.txt')
 # Mock all the modules that are included by lago, so autoimport works as
 # expected with no need to download them (some are not in pip even)
 autodoc_mock_imports = [
+    'enum',
     'guestfs',
     'libvirt',
     'lockfile',
