+++ setup.cfg
@@ -60,6 +60,17 @@ lago.plugins.output =
     json=lago.plugins.output:JSONOutFormatPlugin
     yaml=lago.plugins.output:YAMLOutFormatPlugin
 
+lago.plugins.vm_provider =
+    local_libvirt=lago.vm:LocalLibvirtVM
+    ovirt-node=ovirtlago.virt:NodeVM
+    ovirt-engine=ovirtlago.virt:EngineVM
+    ovirt-host=ovirtlago.virt:HostVM
+
+lago.plugins.vm_service =
+    systemd=lago.service:SystemdService
+    sysvinit=lago.service:SysVInitService
+    systemdcontainer=lago.service:SystemdContainerService
+
 lago.plugins.ovirt.cli =
     collect=ovirtlago.cmd:do_ovirt_collect
     deploy=ovirtlago.cmd:do_deploy
